M1-B9: output device and mixer options in the web player
MixerMode (starplayer-engine, no_std, u32 wire encoding) describes path,
interpolator, output depth, dither and channel count. The host owns the switch:
an eight-arm enum of engine instantiations (float/fixed × nearest/linear ×
mono/stereo) built by a macro, with depth and dither as a post-quantisation
stage using the mixer's own HostSample conversions and seeded TPDF — fixed +
16-bit is byte-identical to the fixed engine's native output. A switch keeps
the same Arc<Module>, seeks to the sounding order, restarts the musical clock
and restores volume, mutes and transport, off the render path.

The page gains an Output panel (actual vs requested sample rate, latencies,
channels, sink, engine mode; rate changes rebuild the AudioContext beside the
live one and resume at the same order; setSinkId where available) and a Mixer
panel (path, interpolation, depth, dither), persisted in localStorage. Chromium
151 honours every rate from 8000 to 96000 exactly. Verified headless in all
three transport modes with the new scenario.

+++ crates/starplayer-host-wasm/src/command.rs
@@ -15,8 +15,8 @@ const CAPACITY_MASK: usize = COMMAND_RING_CAPACITY - 1;
 #[derive(Clone, Copy, Debug, Default, PartialEq, Eq)]
 pub struct WireCommand {
     pub opcode: u8,
-    pub argument: u16,
-    pub extra: u16,
+    pub argument: u32,
+    pub extra: u32,
 }
 
 /// A bounded, allocation-free queue. It is SPSC at the browser boundary: the worklet's
@@ -81,7 +81,7 @@ mod tests {
     fn overflow_is_visible_and_preserves_queued_records() {
         let mut ring = CommandRing::new();
         for index in 0..COMMAND_RING_CAPACITY {
-            assert!(ring.push(WireCommand { opcode: 1, argument: index as u16, extra: 0 }));
+            assert!(ring.push(WireCommand { opcode: 1, argument: index as u32, extra: 0 }));
         }
         assert!(!ring.push(WireCommand::default()));
         assert_eq!(ring.dropped(), 1);
